Update test and docs to prepare for explicit bases
Adding support for dec/bin, which I would like, is really tricky to do nicely when there's an unusual default (hex in this case).
This format requires a base to be given explicitly, but it's still a pretty concise format.

The tests are failing - I'll do the implementation next. TDD!